Moved recursion to the end and removed return statements.
The recursion was moved to the end of the function. Now the function first handles the first indexing level, then calls subsref again for the remaining indexing levels.
This change allows indexing like b(2:3).Value(1) , where .Value returns a double.

The changes to the code mostly concern identation, as return statements were removed and thus code needed to be indented in else statements.