Added support for more than 2 dimensions for horzcat and vertcat, i.e. e = [a b; c d]
These functions now behave the same ways as they do for double variables.
The new implemtation relies on subasgn and does not anymore use am.VertCat(bm) / am.HorzCat(bm). For better performance, it might be a good idea to introduce a special case for 2 dimensions, although the impact should be small when the arrays are large.