Use short-circuiting or operator in IsCompexNet() and IsArrayNet()
In my test case (OSM calibration and deembedding) this change results in a 30%-50% speed improvement of both function.

IsCompexNet() and IsArrayNet() used the element-wise or operator `|`, but isa always returns a scalar. Thus, they could instead also use the logical or operator `||`, which is short-circuiting, i.e. only evaluates the right-hand side if the result is not fully determined by the left-hand side.

I also reordered the statements from most to least likely (in calcualtions related to Vna Tools). However, that order is certainly debatable.

As described on https://de.mathworks.com/help/matlab/ref/logicaloperatorsshortcircuit.html, the element-wise or operator `|` only uses short-circuiting in if and while statements - and it is recommended to use the logical operator `||` instead.