Used a temp variable to save 10% of runtime of size().
I am not 100% sure why this simple change is so usefull.

Changing the type of a variable is expensive and not reccomended in performance-critical code. My guess is that the nested function call used some internal variable that matlab resues at a later point and then requires a type change.